+++ changelog.md
@@ -1,7 +1,7 @@
 # Vocal Raid Assistant
 
 ## 2.5.0
-- Added support for WotLK Classic
+- Added support for Wrath Classic
 
 ## 2.4.4
 - Fixed incorrect spell-ID for Alter Time